+++ pkg/policy/l4_filter_deny_test.go
@@ -94,7 +94,7 @@ func (ds *PolicyTestSuite) TestMergeDenyAllL3(c *C) {
 
 	filter, ok := l4IngressDenyPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(filter.SelectsAllEndpoints(), Equals, true)
@@ -141,7 +141,7 @@ func (ds *PolicyTestSuite) TestMergeDenyAllL3(c *C) {
 
 	filter, ok = l4IngressDenyPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(filter.SelectsAllEndpoints(), Equals, true)
